refactor: split gateway pairing read models
